requirements.txt
Django
djangorestframework
python-dotenv
anthropic